backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Pillow==11.0.0
opencv-python==4.8.1.78
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
Celery==5.3.1
redis==5.0.0
psycopg2-binary==2.9.11
SQLAlchemy==2.0.23